% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsmsr.R
\name{state_mindfulness_scale}
\alias{state_mindfulness_scale}
\title{Calculate State Mindfulness Scale values}
\usage{
state_mindfulness_scale(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Calculate State Mindfulness Scale values
}
\keyword{Mindfulness}
\keyword{Scale}
\keyword{State}
