% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efsmsr.R
\name{state_mindfulness_scale}
\alias{state_mindfulness_scale}
\title{Calculate State Mindfulness Scale (SMS) values}
\usage{
state_mindfulness_scale(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame
}
\description{
Calculates SMS total score, and scores for mind and body subscales (Tanay & Bernstein, 2013).  Expects a data frame
formatted by \code{\link[expfactoryr:process_expfactory_survey]{expfactoryr::process_expfactory_survey}}, containing
responses from a single participant. Participant id \emph{must} be in column \code{p}.
}
\examples{

sms_df <- expand.grid(token = participants$token, survey = 'state-mindfulness-survey-results.json') \%>\%
  rowwise() \%>\%
  do(., expfactoryr::process_expfactory_survey(.$token, paste('data/', .$token, '_finished/', .$survey, sep=''),
    flat=TRUE)) \%>\%
    rename(p = Token)
 process_sms_data <- function(participant, df) {
   return(efsmsr::state_mindfulness_scale(df \%>\% filter(p == participant)))
 }
 sms <- sms_df \%>\%
   rowwise() \%>\%
   do(., process_sms_data(.$p, sms_df)) \%>\%
   arrange(p)

}
\keyword{Mindfulness}
\keyword{Scale}
\keyword{State}
